Moved OpenFileParams to ibisapi.* in order to avoid using FileReader when it is not necessary. Added some functions tpo IbisAPI.
