+++ IbisLib/filereader.h
@@ -25,43 +25,7 @@ class vtkMatrix4x4;
 class ImageObject;
 class PointsObject;
 class IbisAPI;
-
-class OpenFileParams
-{
-
-public:
-
-    struct SingleFileParam
-    {
-        SingleFileParam() : isReference(false), isLabel(false), loadedObject(0), secondaryObject(0), parent(0) {}
-        QString fileName;
-        QString objectName;
-        bool isReference;
-        bool isLabel;  // load image as label image instead of floats.
-        SceneObject * loadedObject;
-        SceneObject * secondaryObject;  // This is a hack to attach the second point object that can be found in PointsObjects
-        SceneObject * parent;
-    };
-    OpenFileParams() : defaultParent(0) {}
-    ~OpenFileParams() {}
-    void AddInputFile( QString filename, QString objectName = QString() )
-    {
-        SingleFileParam p;
-        p.fileName = filename;
-        p.objectName = objectName;
-        filesParams.push_back( p );
-    }
-    void SetAllFileNames( const QStringList & inFiles )
-    {
-        for( int i = 0; i < inFiles.size(); ++i )
-        {
-            AddInputFile( inFiles[i] );
-        }
-    }
-    QList<SingleFileParam> filesParams;
-    QString lastVisitedDir;
-    SceneObject * defaultParent;
-};
+class OpenFileParams;
 
 class FileReader : public QThread
 {
@@ -72,7 +36,7 @@ public:
     FileReader( QObject * parent = 0 );
     ~FileReader();
 
-    void SetParams( OpenFileParams * params ) { m_params = params; }
+    void SetParams( OpenFileParams * params );
     void SetFileNames( QStringList & filenames );  // helper that eventually call SetParams
     void GetReadObjects( QList<SceneObject*> & objects );  // helper
